requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
chart-studio==1.1.0
contextlib2==21.6.0
cycler==0.11.0
dicom2nifti==2.3.4
einops==0.4.1
fonttools==4.31.1
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.2
joblib==1.1.0
kiwisolver==1.4.0
linecache2==1.0.0
Markdown==3.3.6
matplotlib==3.5.1
MedPy==0.4.0
ml-collections==0.1.1
monai==0.6.0
networkx==2.6.3
nibabel==3.2.2
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==9.0.1
plotly==5.6.0
protobuf==3.19.4
ptflops==0.6.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
SimpleITK==2.0.2
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
tenacity==8.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.5.4
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
traceback2==1.4.0
typing_extensions==4.1.1
unittest2==1.1.0
urllib3==1.26.8
webcolors==1.11.1
Werkzeug==2.0.3
zipp==3.7.0
